WIP. Tenant Policy Scope extracted to concern in core
